refactor: Applied feedbacks.
